# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Documentacao da API de E-COMMERCE",
    "version": "1.0.0",
    "title": "ERP E-COMMERCE",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/produto": {
      "get": {
        "tags": [
          "Produto Request"
        ],
        "summary": "Retorna todos os produtos",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/produto"
            }
          }
        }
      },
      "post": {
        "summary": "Cria um produto",
        "tags": [
          "Produto Request"
        ],
        "requestBody": {
          "description": "Objeto do produto que se espera receber",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/produto"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "JSON INVALIDO / ID INVALIDO"
          }
        }
      }
    },
    "/produto/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID do produto que deseja procurar",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Produto Request"
        ],
        "summary": "Retorna o produto pelo ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Id do produto",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/produto"
            }
          },
          "400": {
            "description": "Id Invalido"
          }
        }
      },
      "put": {
        "summary": "Altera um produto pelo id",
        "tags": [
          "Produto Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID do produto é necessário",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "description": "Objeto do produto que se espera receber",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/produto"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "JSON INVALIDO / ID INVALIDO"
          }
        }
      },
      "delete": {
        "summary": "Deleta produto pelo ID",
        "tags": [
          "Produto Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID do produto é necessário",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "ID Inválido"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "id": {
        "properties": {
          "uuid": {
            "type": "string"
          }
        }
      },
...
      "produto": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string",
            "format": "string",
            "example": "Blusa Nike - Preto"
          },
          "preco": {
            "type": "number",
            "format": "number",
            "example": 250.0
          }
        },
        "required": ["desricao","preco"]
      },
...
    }
  }
}